update: modify VolumeController.tsx as index.tsx, modify mute function logic, extract default volume as a constant variable
