+++ src/contexts/VolumeContext.tsx
@@ -3,15 +3,15 @@ import { createContext, useState } from "react";
 interface VolumeContextType {
   volumeLevel: number;
   setVolumeLevel: React.Dispatch<React.SetStateAction<number>>;
-  backtrackVolumeLevel: number;
-  setBacktrackVolumeLevel: React.Dispatch<React.SetStateAction<number>>;
+  preMuteVolumeLevel: number;
+  setPreMuteVolumeLevel: React.Dispatch<React.SetStateAction<number>>;
 }
 
 export const VolumeContext = createContext<VolumeContextType>({
   volumeLevel: 50,
   setVolumeLevel: () => {},
-  backtrackVolumeLevel: 50,
-  setBacktrackVolumeLevel: () => {},
+  preMuteVolumeLevel: 50,
+  setPreMuteVolumeLevel: () => {},
 });
 
 interface Props {
@@ -20,11 +20,11 @@ interface Props {
 
 export const VolumeContextProvider = ({ children }: Props) => {
   const [volumeLevel, setVolumeLevel] = useState<number>(50);
-  const [backtrackVolumeLevel, setBacktrackVolumeLevel] = useState<number>(50);
+  const [preMuteVolumeLevel, setPreMuteVolumeLevel] = useState<number>(50);
 
   return (
     <VolumeContext.Provider
-      value={{ volumeLevel, setVolumeLevel, backtrackVolumeLevel, setBacktrackVolumeLevel }}
+      value={{ volumeLevel, setVolumeLevel, preMuteVolumeLevel, setPreMuteVolumeLevel }}
     >
       {children}
     </VolumeContext.Provider>
